+++ e2e-tests/test/data/transactions.json
@@ -1,3 +1,3 @@
 {
-  "schema": "{\n  \"type\": \"record\",\n  \"name\": \"transaction\",\n  \"namespace\": \"pl.touk\",\n  \"fields\": [\n    {\n      \"name\": \"clientId\",\n      \"type\": \"string\"\n    },\n      {\n      \"name\": \"amount\",\n      \"type\": \"int\"\n    }\n ]\n}\n"
+  "schema": "{\n  \"type\": \"object\",\n  \"properties\": {\n    \"clientId\": {\n      \"type\": \"string\"\n    },\n    \"amount\": {\n      \"type\": \"integer\"\n    }\n  },\n  \"required\": [\n    \"clientId\",\n    \"amount\"\n  ]\n}", "schemaType": "JSON"
 }
